+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Toutes les modifications notables de ce projet seront documentées dans ce fichi
 ## Ajouts :
 - [#327](https://github.com/Refhi/Weda-Helper/pull/327) - Ajout d'une assistance à l'import des documents avec lecture des datamatrix si besoin
 - Amélioration de la navigation par tabulation dans l'import des documents. Alt+S permet désormais de supprimer le document en cours.
+- [#50](https://github.com/Refhi/Weda-Helper/issues/50) - Ajout d'un raccourcis Ctrl+Shift+S pour lancer le scan de documents
 
 # [2.9]
 ## Ajouts :